More WIP - Alternate Level Design Part 1
More WIP on my level design. So far have the first 'room' close to its final completion? The level is now a bit more split up since I added doors that can split up areas into rooms. Still could do with a lot more work on the level design side. But it's coming along. 
Just a note, this is more of an exploration of a possible different level layout, not necessarily the one we are going with... 
That might end up being Tiffany's level. But that will be up to everyone else's opinions. 

Things to note; since my last commit I tweaked some settings on the player mouse - but I have not applied them to the prefab. Take a look at the changes I have made since I think they feel closer to the final design we are going for.